Update README.md, Close #20, and  Close #22
Updated the README.md file to reflect the current state of the template.
Removed references to Flash messages to Close #22,
and changed some of the references to the file structure as it will be
updated to reflect a more modular `React` structure.  Ran markdown
command line tools. Fixed the issue that pointed out the name of the
template was wrong to Close #20.