Clean up scss files and Close #19
Clean up all of the scss files and Close #19 by removing the unused
styles for the flash messages.  Also removed the App.scss (can be added
by devs if they need it). Added an index.scss so that the Bootstrap sass
file is being brought in there to allow for Bootstrap sass variable
overrides.